General Update
In this update I have temporarily removed diagonal movement due to issues with camera scrolling. I have also added land objects for the player to roam around in, when the player reaches an edge they will either be unable to continue in that direction or they will appear on a new section of the example level.